requirements.txt
astroid==2.2.5
dj-database-url==0.5.0
Django==2.1.7
django-haystack==2.8.1
django-heroku==0.3.1
gunicorn==19.9.0
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.7
pylint==2.3.1
pylint-django==2.0.6
pylint-plugin-utils==0.5
pytz==2018.9
six==1.12.0
typed-ast==1.3.1
whitenoise==4.1.2
django-filter==2.1.0
django-widget-tweaks==1.4.3
Whoosh==2.7.4
wrapt==1.11.1
pytest-django==3.4.8
selenium
python-dateutil
